push

+++ ScavengersCST.ps1
@@ -36,6 +36,7 @@ $RepairToolButton_Click = {
   Set-Variable -Name "scstinfo" -Value "$env:temp\SCSTInfo.log"
   Set-Variable -Name "scstrepair" -Value "$env:temp\SCSTRepair.log"
   Set-Variable -Name "uploadtype" -Value "Repair"
+  Set-Variable -Name "dismlog" -Value "$env:TEMP\SCSTDISM.log"
   UsernameInput
   $CloseButton.Enabled = $false
   $DataCollectorButton.Enabled = $false
@@ -48,16 +49,18 @@ $RepairToolButton_Click = {
   UploadtoDiscord -Wait
   Set-Variable -Name "logfiletoupload" -Value "$scstrepair"
   UploadtoDiscord -Wait
+  Write-Host "Now uploading our log files to Discord"
   Cleanup -Wait
+  Write-Host "Now cleaning up the log files"
   $CloseButton.Enabled = $true
   $RepairToolButton.Enabled = $true
   $DataCollectorButton.Enabled = $true
 }
 $AcceptCheckbox_CheckedChanged = {
-  if ($AcceptCheckBox.Checked) { $DataCollectorButton.Visible = $true }
-  else { $DataCollectorButton.Visible = $false }
-  if ($AcceptCheckBox.Checked) { $RepairToolButton.Visible = $true }
-  else { $RepairToolButton.Visible = $false }
+  if ($AcceptCheckBox.Checked) {$DataCollectorButton.Visible = $true}
+  else {$DataCollectorButton.Visible = $false}
+  if ($AcceptCheckBox.Checked) {$RepairToolButton.Visible = $true}
+  else {$RepairToolButton.Visible = $false}
 }
 function ElevatetoAdmin
 {
@@ -73,10 +76,8 @@ function UsernameInput {
   $title = 'Scavengers Community Support Tool'
   $msg = 'Please Enter Your Discord Username Including the Numbers'
   $discordusernameentry = [Microsoft.VisualBasic.Interaction]::InputBox($msg,$title)
-  if (Test-Path $env:TEMP\SCSTDiscord.txt) {
-    Set-Content -Path "$env:TEMP\SCSTDiscord.txt" -Value "$discordusernameentry" -Encoding UTF8 }
-  else {
-    New-Item "$env:TEMP\SCSTDiscord.txt"
+  if (Test-Path $env:TEMP\SCSTDiscord.txt) {Set-Content -Path "$env:TEMP\SCSTDiscord.txt" -Value "$discordusernameentry" -Encoding UTF8}
+  else {New-Item "$env:TEMP\SCSTDiscord.txt"
     Set-Content -Path "$env:TEMP\SCSTDiscord.txt" -Value "$discordusernameentry" -Encoding UTF8 }
 }
 function UploadtoDiscord
@@ -113,23 +114,19 @@ $discordusername Submitted this $uploadtype log file.
 }
 function Cleanup
 {
-  if (Test-Path $env:TEMP\SCSTDiscord.txt) {
-    Remove-Item -Path "$env:temp\SCSTDiscord.txt" }
+  if (Test-Path $env:TEMP\SCSTDiscord.txt) {Remove-Item -Path "$env:temp\SCSTDiscord.txt"}
   else {}
-  if (Test-Path $env:temp\SCSTInfo.log) {
-    Remove-Item -Path "$env:temp\SCSTInfo.log" }
+  if (Test-Path $env:temp\SCSTInfo.log) {Remove-Item -Path "$env:temp\SCSTInfo.log"}
   else {}
-  if (Test-Path $env:temp\SCSTRepair.log) {
-    Remove-Item -Path "$env:temp\SCSTRepair.log" }
+  if (Test-Path $env:temp\SCSTRepair.log) {Remove-Item -Path "$env:temp\SCSTRepair.log"}
   else {}
-  if (Test-Path $env:temp\SCSTDISM.log) {
-    Remove-Item -Path "$env:temp\SCSTDISM.log" }
+  if (Test-Path $env:temp\SCSTDISM.log) {Remove-Item -Path "$env:temp\SCSTDISM.log"}
   else {}
 }
 function CollectInfo
 {
-  if (Test-Path $scstinfo) {}
-  else { New-Item "$env:temp\SCSTInfo.log" }
+  if (Test-Path $scstinfo) {Remove-Item -Path "$scstinfo"}
+  else {New-Item "$env:temp\SCSTInfo.log"}
   Get-ComputerInfo | Select-Object WindowsProductName,WindowsVersion,OsHardwareAbstractionLayer | Out-File -FilePath "$scstinfo" -Encoding utf8 -Force
   Get-CimInstance -ClassName CIM_Processor | Select-Object -Property Name,MaxClockSpeed,SocketDesignation,Manufacturer | Out-File -FilePath "$scstinfo" -Encoding utf8 -Append
   Get-CimInstance -ClassName CIM_VideoController | Format-Table -AutoSize -Property Name,CurrentHorizontalResolution,CurrentVerticalResolution,CurrentRefreshRate,@{ Name = "AdapterRamGB"; Expression = { [int]($_.AdapterRam / 1GB) } },DriverDate,DriverVersion | Out-File -FilePath "$scstinfo" -Encoding utf8 -Append
@@ -139,38 +136,44 @@ function CollectInfo
 }
 function Repair
 {
-  $dismlog = "$env:TEMP\SCSTDISM.log"
   if (Test-Path "$dismlog") {}
   else { New-Item "$dismlog" }
-  #ElevatetoAdmin
-  Start-Process -FilePath "$env:SystemRoot\System32\sfc.exe" -ArgumentList "/scannow" -Wait
-  Start-Process -FilePath "$env:SystemRoot\System32\Dism.exe" -ArgumentList "/Online /Cleanup-Image /RestoreHealth /LogPath:$dismlog" -Wait
+  Write-Host "Now running the Windows System File Checker"
+  Start-Process -FilePath "$env:SystemRoot\System32\sfc.exe" -ArgumentList "/scannow" -Wait -NoNewWindow
+  Write-Host "Now running the Windows DISM Tool"
+  Start-Process -FilePath "$env:SystemRoot\System32\Dism.exe" -ArgumentList "/Online /Cleanup-Image /RestoreHealth /LogPath:$dismlog" -Wait -NoNewWindow
   $sourcedirectx = "https://download.microsoft.com/download/8/4/A/84A35BF1-DAFE-4AE8-82AF-AD2AE20B6B14/directx_Jun2010_redist.exe"
   $destinationdirectx = "$env:TEMP\directx_Jun2010_redist.exe"
+  Write-Host "Now downloading the latest DirectX Redistributable From Microsoft"
   Invoke-WebRequest -Uri $sourcedirectx -OutFile "$destinationdirectx"
   $sourcevcredisx86 = "https://aka.ms/vs/16/release/vc_redist.x86.exe"
   $destinationvcredisx86 = "$env:TEMP\vc_redist.x86.exe"
+  Write-Host "Now downloading the latest Visual C+++ 32bit Redistributable From Microsoft"
   Invoke-WebRequest -Uri $sourcevcredisx86 -OutFile "$destinationvcredisx86"
   $sourcevcredisx64 = "https://aka.ms/vs/16/release/vc_redist.x64.exe"
   $destinationvcredisx64 = "$env:TEMP\vc_redist.x64.exe"
+  Write-Host "Now downloading the latest Visual C+++ 64bit Redistributable From Microsoft"
   Invoke-WebRequest -Uri $sourcevcredisx64 -OutFile "$destinationvcredisx64"
+  Write-Host "Now unpacking and installing the DirectX Redistributable"
   $dxsetup = "$env:TEMP\DirectXInstaller\DXSETUP.exe"
-  if (Test-Path -Path "$env:TEMP\DirectXInstaller") {
-    Remove-Item -Path "$env:TEMP\DirectXInstaller" -Recurse
+  if (Test-Path -Path "$env:TEMP\DirectXInstaller") {Remove-Item -Path "$env:TEMP\DirectXInstaller" -Recurse
     Start-Process -FilePath "$destinationdirectx" -ArgumentList "/Q /T:$env:TEMP\DirectXInstaller\" -Wait
-    Start-Process -FilePath "$dxsetup" -ArgumentList "/silent" -Wait }
+    Start-Process -FilePath "$dxsetup" -ArgumentList "/silent" -Wait}
   else { Start-Process -FilePath "$destinationdirectx" -ArgumentList "/Q /T:$env:TEMP\DirectXInstaller\" -Wait
     Start-Process -FilePath "$dxsetup" -ArgumentList "/silent" -Wait }
+  Write-Host "Now installing the Visual C+++ 32bit Redistributable"
   Start-Process -FilePath "$destinationvcredisx86" -ArgumentList "/install /quiet /norestart" -Wait
+  Write-Host "Now installing the Visual C+++ 64bit Redistributable"
   Start-Process -FilePath "$destinationvcredisx64" -ArgumentList "/install /quiet /norestart" -Wait
+  Write-Host "Now deleting all of the installers we just downloaded and ran"
   Remove-Item -Path "$destinationdirectx"
   Remove-Item -Path "$env:TEMP\DirectXInstaller" -Recurse
   Remove-Item -Path "$destinationvcredisx86"
   Remove-Item -Path "$destinationvcredisx64"
-  if (Test-Path $env:TEMP\SCSTRepair.log) {
-    Remove-Item -Path "$env:TEMP\SCSTRepair.log" }
-  else { New-Item "$env:TEMP\SCSTRepair.log" }
+  if (Test-Path $env:TEMP\SCSTRepair.log) {Remove-Item -Path "$env:TEMP\SCSTRepair.log"}
+  else { New-Item "$env:TEMP\SCSTRepair.log"}
   $scstrepair = "$env:TEMP\SCSTRepair.log"
+  Write-Host "Now building our log file"
   Add-Content -Path "$scstrepair" -Value "`r`nSFC LOG BEGINS HERE" -Encoding utf8
   $sr = Get-Content c:\windows\Logs\CBS\CBS.log | Where-Object { $_.Contains("[SR]") } | Select-Object -Property @{ Name = "LastCheckDate"; Expression = { $_.substring(0,10) } } -Last 1
   Get-Content c:\windows\Logs\CBS\CBS.log | Where-Object { $_.Contains("[SR]") -and $_.Contains($sr.lastcheckdate) } | Select-String -NotMatch "Verify complete","Verifying","Beginning Verify and Repair" | Out-File -FilePath "$scstrepair" -Encoding utf8 -Append
